Release v4.0.0 (#309)
* update version and changelog

* update recent main updates

* fix to removing old zip file

+++ bin/archive.sh
@@ -7,7 +7,9 @@ copydest=$1
 cd dist/chrome;
 version=`cat manifest.json | jq -r '.version' | sed 's/\./-/g'`
 zipfile="aesr-chrome-$version.zip"
-\rm $zipfile
+if [ -e $zipfile ]; then
+  \rm $zipfile
+fi
 zip -r $zipfile \
   manifest.json *.html icons/ js/ 
 echo "archived: chrome/$zipfile"
@@ -20,7 +22,9 @@ echo "----"
 cd ../firefox;
 version=`cat manifest.json | jq -r '.version' | sed 's/\./-/g'`
 zipfile="aesr-firefox-$version.zip"
-\rm $zipfile
+if [ -e $zipfile ]; then
+  \rm $zipfile
+fi
 zip -r $zipfile \
   manifest.json *.html icons/ js/
